Fix Render Bug
For some reason, the code would render more cells than it should have. For example, if there was a water level of 1m, the app would render cells with HAND values of 2 or less. I used an offset of 0.97 for the renders  to solve this issue. Again, I don't know the origin of this bug.